refactoring main.go
